+++ frontend/src/components/index.ts
@@ -4,8 +4,8 @@ export { default as TopBackground } from "./common/TopBackground";
 export { default as TopBar } from "./common/TopBar";
 export { default as Title } from "./register/Title";
 export { default as ThemeButton } from "./common/ThemeButton";
-export { default as InputLabel } from "./register/InputLabel";
-export { default as SubmitInput } from "./register/SubmitInput";
-export { default as TagInputLabel } from "./register/TagInputLabel";
-export { default as ProfileImgUploadModal } from "./register/ProfileImgUploadModal";
+export { default as InputLabel } from "./common/InputLabel";
+export { default as SubmitInput } from "./common/SubmitInput";
+export { default as TagInputLabel } from "./profile/TagInputLabel";
+export { default as ProfileForm } from "./profile/ProfileForm";
 export * from "./main/ProfileBox";
